+++ src/zabbix_server/taskmanager/taskmanager.c
@@ -1089,7 +1089,7 @@ static int	tm_process_tasks(zbx_ipc_async_socket_t *rtc, int now)
 
 					zabbix_log(LOG_LEVEL_WARNING, "%s", error);
 					task = zbx_tm_task_create(0, ZBX_TM_TASK_REMOTE_COMMAND_RESULT,
-							ZBX_TM_STATUS_NEW, time(NULL), 0, 0);
+							ZBX_TM_STATUS_NEW, (int)time(NULL), 0, 0);
 					task->data = zbx_tm_remote_command_result_create(taskid, FAIL, error);
 					zbx_tm_save_task(task);
 					zbx_tm_task_free(task);
@@ -1139,7 +1139,7 @@ static int	tm_process_tasks(zbx_ipc_async_socket_t *rtc, int now)
 
 						zabbix_log(LOG_LEVEL_WARNING, "%s", error);
 						task = zbx_tm_task_create(0, ZBX_TM_TASK_DATA_RESULT, ZBX_TM_STATUS_NEW,
-								time(NULL), 0, 0);
+								(int)time(NULL), 0, 0);
 						task->data= zbx_tm_data_result_create(taskid, FAIL, error);
 						zbx_tm_save_task(task);
 						zbx_tm_task_free(task);
